Now have chains displaying when you get a combo
I had to redesign how chains work a bit. When a fall detects a combo, it
passes the amount of chains it is currently keeping track of to the Combo.
Likewise, a Combo passes the amount of chains it is keeping track of to the
Fall.
When a combo occurs and the current chain counter is >= 2 for that Combo,
then a new Chain is added to the chainEvents list in Grid. A Chain holds the
row and column of the combo that the chain occured at.
Chain::display() displays the chain at the coordinates that it occured at.
Grid::display() loops through chainEvents and calls each Chain's display
function.
Grid::composeFrame() loops through chainEvents and deletes ones that have
expired.