Added gameover state and screen to grid
When the player loses, the grid goes into a gameover state. It will start a
timer. While the grid is in the gameover state, the player cannot do any
input. When the timer elapses, it goes back to the main menu. Right now this
all looks really shitty, but it gets the job done. However, I'm not sure how
to disable buttons from working while grid is in a gameover state.